Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/webappapis/dynamic-markup-insertion/document-write/module-tla-immediate-promise-expected.txt

CONSOLE MESSAGE: SyntaxError: Unexpected keyword 'new'

Harness Error (TIMEOUT), message = null

TIMEOUT document.write in an imported module Test timed out
